+++ packages/components/input/src/InputTrailingAddon.tsx
@@ -6,8 +6,8 @@ import { useInputGroup } from './InputGroupContext'
 
 export type InputTrailingAddonProps = InputAddonProps
 
-export const InputTrailingAddon = forwardRef<HTMLDivElement, InputTrailingAddonProps>(
-  ({ asChild = false, className, ...others }, ref) => {
+const Root = forwardRef<HTMLDivElement, InputTrailingAddonProps>(
+  ({ className, ...others }, ref) => {
     const { disabled, readOnly } = useInputGroup()
     const isInactive = disabled || readOnly
 
@@ -15,7 +15,6 @@ export const InputTrailingAddon = forwardRef<HTMLDivElement, InputTrailingAddonP
       <div className={cx('rounded-r-lg', isInactive ? 'bg-on-surface/dim-5' : null)}>
         <InputAddon
           ref={ref}
-          asChild={asChild}
           className={cx(className, '!rounded-l-none rounded-r-lg ml-[-1px]')}
           {...others}
         />
@@ -23,5 +22,9 @@ export const InputTrailingAddon = forwardRef<HTMLDivElement, InputTrailingAddonP
     )
   }
 )
+
+export const InputTrailingAddon = Object.assign(Root, {
+  id: 'TrailingAddon',
+})
 
 InputTrailingAddon.displayName = 'InputGroup.TrailingAddon'