+++ packages/components/textarea/src/Textarea.tsx
@@ -4,14 +4,18 @@ import { ComponentPropsWithoutRef, forwardRef, PropsWithChildren } from 'react'
 
 export type TextareaProps = ComponentPropsWithoutRef<'textarea'>
 
-export const Textarea = forwardRef<HTMLTextAreaElement, PropsWithChildren<TextareaProps>>(
+const Root = forwardRef<HTMLTextAreaElement, PropsWithChildren<TextareaProps>>(
   ({ className, disabled, rows = 1, ...others }, ref) => {
     return (
-      <Input className={cx(className, 'py-md resize-y')} disabled={disabled} asChild>
+      <Input className={cx(className, 'py-[var(--sz-10)] resize-y')} disabled={disabled} asChild>
         <textarea ref={ref} rows={rows} {...others} />
       </Input>
     )
   }
 )
+
+export const Textarea = Object.assign(Root, {
+  id: Input.id,
+})
 
 Textarea.displayName = 'Textarea'